changes at manhattan & ANN

+++ manhattan.h
@@ -1,1 +1,1 @@
-int manhattan(unsigned char* , unsigned char* , int );
+double manhattan(unsigned char* , unsigned char* , int );
